refactor(ui): put user badges into separate files
